add an benchmark that tests the impact of the error variant
