+++ martin/Cargo.toml
@@ -76,7 +76,7 @@ path = "src/bin/martin-cp.rs"
 required-features = ["mbtiles"]
 
 [[bench]]
-name = "bench"
+name = "sources"
 harness = false
 
 [features]
